+++ server/tests-py/test_metadata.py
@@ -262,3 +262,14 @@ class TestMetadataOrder:
         # `resource_version` should be unchanged
         assert export_resp['resource_version'] == export_resp_1['resource_version']
 
+
+@pytest.mark.parametrize("backend", ['citus', 'mssql', 'postgres'])
+@usefixtures('per_class_tests_db_state')
+class TestSetTableCustomization:
+
+    @classmethod
+    def dir(cls):
+        return "queries/v1/metadata"
+
+    def test_set_table_customization(self, hge_ctx):
+        check_query_f(hge_ctx, self.dir() + hge_ctx.backend_suffix('/set_table_customization') + '.yaml')