+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 (Add entries below in the order of server, console, cli, docs, others)
 
 - server: make improvements in the `livequery-poller-log`
+- server: Backends Citus, MSSQL, and BigQuery now all support the `set_table_customization` operation.
 - cli: `metadata diff` will now only show the differences in metadata. old behaviour is avialble behind a flag (`--type unified-common`) (#5487)
 
 ## v2.0.0-beta.2